[Finish]3-9

+++ app/controllers/books_controller.rb
@@ -1,6 +1,5 @@
 class BooksController < ApplicationController
   before_action :authenticate_user!
-
   def index
     @book = Book.new
     @books = Book.all.order(:id)
@@ -21,14 +20,24 @@ class BooksController < ApplicationController
 
   def show
     @book = Book.find(params[:id])
+    @new_book = Book.new
   end
 
   def edit
     @book = Book.find(params[:id])
+    unless current_user.id == @book.user_id
+      redirect_to books_path
+      return
+    end
   end
 
   def update
     @book = Book.find(params[:id])
+    unless current_user.id == @book.user_id
+      redirect_to book_path(@book)
+      return
+    end
+
     if @book.update(book_params)
       flash[:notice] = "successfully updated"
       redirect_to book_path(@book)
@@ -39,6 +48,11 @@ class BooksController < ApplicationController
 
   def destroy
     book = Book.find(params[:id])
+    unless current_user.id == book.user_id
+      redirect_to book_path(book)
+      return
+    end
+
     book.destroy
     flash[:notice] = "successfully deleted"
     redirect_to books_path
@@ -49,4 +63,5 @@ class BooksController < ApplicationController
   def book_params
     params.require(:book).permit(:title, :body)
   end
+
 end
